LayoutTests/canvas/philip/tests/2d.imageData.put.dirty.negative-expected.txt

< [index] >
2d.imageData.put.dirty.negative
putImageData() handles negative-sized dirty rectangles correctly
References: 2d.imageData.put.normal
Actual output:
Expected output:

Passed
